Fix tab insertion when `tabAutocomplete` is true
Given this C code (should not matter):

```c
// replace '@' with hashtag
@define _INPUT(X, ...) 				                \
    enum var blah;\
    //            ^--- here
```

You can't insert a tab between the semicolon and the backslash if
`tabAutocomplete` is enabled.

This was caused because `GetWord()` (`./internal/buffer/autocomplete.go`) returns
an empty array of bytes ([]byte{}) when it fails, but the integer returned
changes depending on which condition failed.

By checking the length of the array of bytes, it's resolved.